Multiple statements working (#127)
This lets you have multiple queries in the editor so you can do things
like this.

<img width="1916" alt="image"
src="https://github.com/user-attachments/assets/8b6f3443-0e5f-4b36-b6a1-f87dd3a172bc">

~Its still a bit messy~ (ive cleaned it up). Ive had to play with how
the execution will work. Initially tried running a `Statement` but since
in the editor we want to store the raw sql with each execution it was
easier to just stick with that for now.

Playing with this, after the additions for the CLI features, definitely
makes it clear well need to figure out the appropriate execution
abstractions.