Feat/sql pagination 2 (#157)
(https://github.com/datafusion-contrib/datafusion-dft/pull/141) was
becoming hard for me to reason about because FlightSQL pagination has to
be handled differently than SQL tab pagination (because on the SQL tab
we can control batch size directly on the context but we cant do that
with FlightSQL). So I would like to split pagination on each of those
tabs into their own PRs to keep them more focused and easier to review /
reason about / etc.

So this PR makes pagination work and adds integration tests for testing
pagination. There is still room for improvement in the end to end
testing (right now we test some of the implementation details) but at
least we have some coverage.