Test DDL (#161)
This adds test for running `AppEvent::ExecuteDDL` and then running query
to make sure the DDL statements were correctly processed.

Unfortunately, this isnt a proper end to end test like I was hoping. It
was surprisingly difficult to get the app event loop to run in the test
which is required to get that to work. So, we simply execute a query
directly on the context instead.