Start setting up flightsql pagination (#163)
This adds basic pagination for FlightSQL. Only basic because for this
initial implementation it only paginates over records in a record batch.
It does not yet handle paginating over multiple record batches. The
structure is in place for that (with `take_record_batches` but
implementation details need to be figured out.